Improved claim system

+++ api/cmd/admin/finance/_financeCallbacks.ts
@@ -4,7 +4,6 @@ import { loadFunction, removeInLineButton } from '../../../app/_telefunctions';
 import {
   chat,
   dbMessaging,
-  gSheetDB,
   team,
 } from '../../../database_mongoDB/functions/_index';
 import { Database } from '../../../database_mongoDB/_db-init';
@@ -17,6 +16,8 @@ import { gsheet } from '../../../functions/_initialise';
 import { initial } from '../../../models/_SessionData';
 import { v4 as uuidv4 } from 'uuid';
 import { DateTime } from 'luxon';
+import { searchRowNo } from '../../../gsheets/_gsheet_functions';
+import { GoogleSpreadsheetRow } from 'google-spreadsheet';
 
 export const adminFinance = (bot: Bot<BotContext>) => {
   //Finance Team Management
@@ -24,18 +25,35 @@ export const adminFinance = (bot: Bot<BotContext>) => {
 
   //Fund Management
   bot.callbackQuery('fundManagement', loadFunction, fundMenu);
-  bot.callbackQuery('addOffering', addOffering);
-  bot.callbackQuery(/^addOffering-/, addOfferingAmount);
-  bot.callbackQuery('rmOffering', deleteOffering);
+  bot.callbackQuery('addOffering', loadFunction, addOffering);
+  bot.callbackQuery(/^addOffering-/, loadFunction, addOfferingAmount);
+  bot.callbackQuery('rmOffering', loadFunction, deleteOffering);
 
   //Reimbursement Management
   bot.callbackQuery('reimbursementManagement', loadFunction, reimbursementMenu);
-  bot.callbackQuery('approveReimbursement', approveReimbursement);
-  bot.callbackQuery(/^approveReimbursement-/, approveReimbursementFunction);
-  bot.callbackQuery('rejectReimbursement', rejectReimbursement);
-  bot.callbackQuery(/^rejectReimbursement-/, rejectReimbursementFunction);
-  bot.callbackQuery('completedReimbursement', completedReimbursement);
-  bot.callbackQuery(/^completedReimbursement-/, completedReimbursementWitness);
+  bot.callbackQuery('viewAllClaims', loadFunction, viewAllClaims);
+  bot.callbackQuery('approveReimbursement', loadFunction, approveReimbursement);
+  bot.callbackQuery(
+    /^approveReimbursement-/,
+    loadFunction,
+    approveReimbursementFunction
+  );
+  bot.callbackQuery('rejectReimbursement', loadFunction, rejectReimbursement);
+  bot.callbackQuery(
+    /^rejectReimbursement-/,
+    loadFunction,
+    rejectReimbursementFunction
+  );
+  bot.callbackQuery(
+    'completedReimbursement',
+    loadFunction,
+    completedReimbursement
+  );
+  bot.callbackQuery(
+    /^completedReimbursement-/,
+    loadFunction,
+    completedReimbursementWitness
+  );
   bot.callbackQuery(
     /^completedReimbursementWitness-/,
     completedReimbursementAmount
@@ -44,10 +62,20 @@ export const adminFinance = (bot: Bot<BotContext>) => {
     /^completedReimbursementAmount-/,
     completedReimbursementFunction
   );
+  bot.callbackQuery('deleteReimbursements', loadFunction, deleteReimbursements);
+  bot.callbackQuery(
+    /^deleteReimbursements-/,
+    loadFunction,
+    deleteReimbursementFunction
+  );
 
   //Change Finance Password
-  bot.callbackQuery('changeFinancePassword', changeFinancePassword);
-  bot.callbackQuery(/^changePassword/, cfmPassword);
+  bot.callbackQuery(
+    'changeFinancePassword',
+    loadFunction,
+    changeFinancePassword
+  );
+  bot.callbackQuery(/^changePassword/, loadFunction, cfmPassword);
 
   //Change Finance Chat
   chat.chooseChat(bot, 'Finance');
@@ -129,6 +157,12 @@ const reimbursementMenu = async (ctx: CallbackQueryContext<BotContext>) => {
   const inlineKeyboard = new InlineKeyboard([
     [
       {
+        text: 'View All Claims',
+        callback_data: 'viewAllClaims',
+      },
+    ],
+    [
+      {
         text: 'Approve Reimbursement',
         callback_data: 'approveReimbursement',
       },
@@ -145,6 +179,12 @@ const reimbursementMenu = async (ctx: CallbackQueryContext<BotContext>) => {
         callback_data: 'completedReimbursement',
       },
     ],
+    [
+      {
+        text: 'Delete Reimbursements',
+        callback_data: 'deleteReimbursements',
+      },
+    ],
   ]);
   await ctx.reply(
     `<b>All Statuses:</b>\nPending Approval 🟠\nPending Reimbursement 🟠\nCompleted ✅\n\nTo Be Reimbursed Amount: $${reimburse}\n\nTotal Claims: ${totalClaims}\nAwaiting Approval: ${awaitingApprovedClaims}\nAwaiitng Reimbursement: ${awaitingReimbursementClaims}\nCompleted: ${completedClaims}`,
@@ -155,6 +195,24 @@ const reimbursementMenu = async (ctx: CallbackQueryContext<BotContext>) => {
   );
 };
 
+const viewAllClaims = async (ctx: CallbackQueryContext<BotContext>) => {
+  removeInLineButton(ctx);
+  const approvalClaims = await Database.getMongoRepository(Claims).find({
+    status: 'Pending Approval 🟠',
+  });
+  const reimbursementClaims = await Database.getMongoRepository(Claims).find({
+    status: 'Pending Reimbursement 🟠',
+  });
+  const claims = approvalClaims.concat(reimbursementClaims);
+  if (claims.length === 0) {
+    await ctx.reply('No claims to view');
+    return;
+  }
+  claims.map(async (n) => {
+    await ctx.reply(n.msg, { parse_mode: 'HTML' });
+  });
+};
+
 const approveReimbursement = async (ctx: CallbackQueryContext<BotContext>) => {
   removeInLineButton(ctx);
   const claims = await Database.getMongoRepository(Claims).find({
@@ -163,7 +221,7 @@ const approveReimbursement = async (ctx: CallbackQueryContext<BotContext>) => {
   const inlineKeyboard = new InlineKeyboard(
     claims.map((n) => [
       {
-        text: n.description,
+        text: n.description + ' ( ' + n.name + '  ) - ' + n.date,
         callback_data: `approveReimbursement-${n.claimid}`,
       },
     ])
@@ -177,48 +235,84 @@ const approveReimbursementFunction = async (
   ctx: CallbackQueryContext<BotContext>
 ) => {
   removeInLineButton(ctx);
-  const callback = parseInt(
-    ctx.update.callback_query.data.substring('approveReimbursement-'.length)
+  const callback = ctx.update.callback_query.data.substring(
+    'approveReimbursement-'.length
   );
+
+  const financeGSheet = await gsheet('finance');
+  const claimsSheet = financeGSheet.sheetsByTitle['Claims'];
 
   const claim = await Database.getMongoRepository(Claims).findOneBy({
     claimid: callback,
   });
+  const name = await Database.getMongoRepository(Names).findOneBy({
+    teleUser: ctx.update.callback_query.from.username,
+  });
+  await claimsSheet.loadCells();
 
-  if (!claim) {
+  const claimRowNo = await searchRowNo(
+    callback,
+    claimsSheet,
+    'A',
+    2,
+    claimsSheet.rowCount
+  );
+  const claimRows = await claimsSheet.getRows({});
+  let row: GoogleSpreadsheetRow | null = null;
+
+  if (!claim || !name) {
     await ctx.reply('Invalid Claim');
     return;
   }
+  if (claimRowNo != -1 && claimRowNo) {
+    row = claimRows[claimRowNo - 2];
+  } else {
+    await ctx.reply('Invalid Claim');
+    return;
+  }
+
   const user = claim.name;
   const amount = claim.amount;
   const reason = claim.description;
-  const claimid = claim.claimid;
   const status = 'Pending Reimbursement 🟠';
   const formattedDate = claim.date;
+  const claimMsg = `Claim submitted by\n${user}\n${formattedDate}\n\n<b>${status}</b>\n\nAmount: $${amount}\nDescription: ${reason}`;
 
-  claim.status = 'Pending Reimbursement 🟠';
-  await Database.getMongoRepository(Claims).save(claim);
-  await ctx.api.editMessageCaption(
-    process.env.LG_FINANCE_CLAIM || '',
-    callback,
-    {
-      caption: `Claim submitted by\n${user}\n${formattedDate}\nClaim ID: ${claimid}\n\n<b>${status}</b>\n\nAmount: $${amount}\nDescription: ${reason}`,
-      parse_mode: 'HTML',
-    }
-  );
+  claim.status = status;
+  claim.msg = claimMsg;
+  const savedDBEntry = await Database.getMongoRepository(Claims).save(claim);
+
+  if (!savedDBEntry) {
+    await ctx.reply('Invalid Claim');
+    return;
+  }
+
+  if (row) {
+    row.set('Status', status);
+    row.set('Approved by', name.nameText);
+    row.save();
+  } else {
+    await ctx.reply('Invalid Claim');
+    return;
+  }
+
   await ctx.reply('Claim Approved');
   ctx.session = initial();
 };
 
 const rejectReimbursement = async (ctx: CallbackQueryContext<BotContext>) => {
   removeInLineButton(ctx);
-  const claims = await Database.getMongoRepository(Claims).find({
+  const approvedClaims = await Database.getMongoRepository(Claims).find({
     status: 'Pending Approval 🟠',
   });
+  const reimbursementClaims = await Database.getMongoRepository(Claims).find({
+    status: 'Pending Reimbursement 🟠',
+  });
+  const claims = approvedClaims.concat(reimbursementClaims);
   const inlineKeyboard = new InlineKeyboard(
     claims.map((n) => [
       {
-        text: n.description,
+        text: n.description + ' ( ' + n.name + '  ) - ' + n.date,
         callback_data: `rejectReimbursement-${n.claimid}`,
       },
     ])
@@ -232,9 +326,22 @@ const rejectReimbursementFunction = async (
   ctx: CallbackQueryContext<BotContext>
 ) => {
   removeInLineButton(ctx);
-  const callback = parseInt(
-    ctx.update.callback_query.data.substring('rejectReimbursement-'.length)
+  const callback = ctx.update.callback_query.data.substring(
+    'rejectReimbursement-'.length
   );
+  const financeGSheet = await gsheet('finance');
+  const claimsSheet = financeGSheet.sheetsByTitle['Claims'];
+  await claimsSheet.loadCells();
+  const claimRowNo = await searchRowNo(
+    callback,
+    claimsSheet,
+    'A',
+    2,
+    claimsSheet.rowCount
+  );
+
+  const claimRows = await claimsSheet.getRows({});
+  let row: GoogleSpreadsheetRow | null = null;
 
   const claim = await Database.getMongoRepository(Claims).findOneBy({
     claimid: callback,
@@ -244,19 +351,32 @@ const rejectReimbursementFunction = async (
     await ctx.reply('Invalid Claim');
     return;
   }
+  if (claimRowNo != -1 && claimRowNo) {
+    row = claimRows[claimRowNo - 2];
+  } else {
+    await ctx.reply('Invalid Claim');
+    return;
+  }
+
   const user = await Database.getMongoRepository(Names).findOneBy({
     nameText: claim.name,
   });
+
   if (!user) {
     await ctx.reply('Invalid User');
     return;
   }
+  if (!row) {
+    await ctx.reply('Invalid Claim');
+    return;
+  }
+
   const teleUser = user.teleUser;
   const desc = claim.description;
   await Database.getMongoRepository(Claims).deleteOne({
     claimid: callback,
   });
-  await ctx.api.deleteMessage(process.env.LG_FINANCE_CLAIM || '', callback);
+  await row.delete();
   await dbMessaging.sendMessageUser(
     teleUser,
     `Your claim (${desc}) has been rejected. Please contact the finance personnel if you have any queries!`,
@@ -276,7 +396,7 @@ const completedReimbursement = async (
   const inlineKeyboard = new InlineKeyboard(
     claims.map((n) => [
       {
-        text: n.description,
+        text: n.description + ' ( ' + n.name + '  ) - ' + n.date,
         callback_data: `completedReimbursement-${n.claimid}`,
       },
     ])
@@ -285,12 +405,13 @@ const completedReimbursement = async (
     reply_markup: inlineKeyboard,
   });
 };
+
 const completedReimbursementWitness = async (
   ctx: CallbackQueryContext<BotContext>
 ) => {
   removeInLineButton(ctx);
-  const callback = parseInt(
-    ctx.update.callback_query.data.substring('completedReimbursement-'.length)
+  const callback = ctx.update.callback_query.data.substring(
+    'completedReimbursement-'.length
   );
   const claim = await Database.getMongoRepository(Claims).findOneBy({
     claimid: callback,
@@ -299,7 +420,7 @@ const completedReimbursementWitness = async (
     await ctx.reply('Invalid Claim');
     return;
   }
-  ctx.session.chatId = callback;
+  ctx.session.claimId = callback;
   const names = await Database.getRepository(Names).find();
   const inlineKeyboard = new InlineKeyboard(
     names.map((n) => [
@@ -321,7 +442,7 @@ const completedReimbursementAmount = async (
   const callback = ctx.update.callback_query.data.substring(
     'completedReimbursementWitness-'.length
   );
-  const claimid = ctx.session.chatId;
+  const claimid = ctx.session.claimId;
   const witness = await Database.getRepository(Names).findOneBy({
     teleUser: callback,
   });
@@ -365,7 +486,7 @@ const completedReimbursementFunction = async (
   const callback = ctx.update.callback_query.data.substring(
     'completedReimbursementAmount-'.length
   );
-  const claimid = ctx.session.chatId;
+  const claimid = ctx.session.claimId;
   const witness = ctx.session.reminderUser;
   const user = ctx.callbackQuery.from.username;
   if (claimid && user && witness) {
@@ -383,10 +504,44 @@ const completedReimbursementFunction = async (
       const financeGSheet = await gsheet('finance');
       const reimbursementSheet = financeGSheet.sheetsByTitle['Reimbursement'];
       const allRecordSheet = financeGSheet.sheetsByTitle['All Records'];
+      const claimsSheet = financeGSheet.sheetsByTitle['Claims'];
+      await claimsSheet.loadCells();
+      const claimRowNo = await searchRowNo(
+        claimid,
+        claimsSheet,
+        'A',
+        2,
+        claimsSheet.rowCount
+      );
+
+      const claimRows = await claimsSheet.getRows({});
+      let row: GoogleSpreadsheetRow | null = null;
+      if (claimRowNo != -1) {
+        row = claimRows[claimRowNo - 2];
+      } else {
+        await ctx.reply('Invalid Claim!');
+        return;
+      }
       const uuid = uuidv4();
       const dateTime = DateTime.now()
         .setZone('Asia/Singapore')
         .toFormat('dd/mm/yyyy hh:mm:ss');
+      const claimMsg = `Claim submitted by\n${claim.name}\n${claim.date}\nClaim ID: ${claim.claimid}\n\n<b>Completed ✅</b>\n\nAmount: $${claim.amount}\nDescription: ${claim.description}`;
+      claim.msg = claimMsg;
+      claim.status = 'Completed ✅';
+      const savedDBEntry =
+        await Database.getMongoRepository(Claims).save(claim);
+      if (!savedDBEntry) {
+        await ctx.reply('Something went wrong! Pls try again!');
+        return;
+      }
+      if (row) {
+        row.set('Status', 'Completed ✅');
+        row.save();
+      } else {
+        await ctx.reply('Something went wrong! Pls try again!');
+        return;
+      }
       await reimbursementSheet.addRow({
         'Transaction ID': uuid,
         'Claim ID': claimid,
@@ -395,6 +550,7 @@ const completedReimbursementFunction = async (
         Amount: -claim.amount,
         Description: claim.description,
         'Approved by': userDoc.nameText,
+        Claimee: claim.name,
         Witness: witness,
       });
       await allRecordSheet.addRow({
@@ -402,22 +558,12 @@ const completedReimbursementFunction = async (
         'Claim ID': claimid,
         Timestamp: dateTime,
         Date: claim.date,
+        Type: 'Reimbursement',
         Amount: -claim.amount,
         Description: claim.description,
         'Approved by': userDoc.nameText,
+        'Claimed by': claim.name,
         Witness: witness,
-      });
-      await ctx.api.editMessageCaption(
-        process.env.LG_FINANCE_CLAIM || '',
-        claimid,
-        {
-          caption: `Claim submitted by\n${claim.name}\n${claim.date}\nClaim ID: ${claim.claimid}\n\n<b>Completed ✅</b>\n\nAmount: $${claim.amount}\nDescription: ${claim.description}`,
-          parse_mode: 'HTML',
-        }
-      );
-
-      await Database.getMongoRepository(Claims).deleteOne({
-        claimid: claimid,
       });
       await ctx.reply('Claim Completed');
       financeGSheet.resetLocalCache();
@@ -429,6 +575,108 @@ const completedReimbursementFunction = async (
       ctx.session.botOnType = 16;
     }
   }
+};
+
+const deleteReimbursements = async (ctx: CallbackQueryContext<BotContext>) => {
+  removeInLineButton(ctx);
+  const claims = await Database.getMongoRepository(Claims).find({
+    status: 'Completed ✅',
+  });
+  const inlineKeyboard = new InlineKeyboard(
+    claims.map((n) => [
+      {
+        text: n.description + ' ( ' + n.name + '  ) - ' + n.date,
+        callback_data: `deleteReimbursements-${n.claimid}`,
+      },
+    ])
+  );
+  await ctx.reply('Choose claim to delete:', {
+    reply_markup: inlineKeyboard,
+  });
+};
+
+const deleteReimbursementFunction = async (
+  ctx: CallbackQueryContext<BotContext>
+) => {
+  removeInLineButton(ctx);
+  const callback = ctx.update.callback_query.data.substring(
+    'deleteReimbursements-'.length
+  );
+
+  const claim = await Database.getMongoRepository(Claims).findOneBy({
+    claimid: callback,
+  });
+  if (!claim) {
+    await ctx.reply('Invalid Claim');
+    return;
+  }
+
+  const financeGSheet = await gsheet('finance');
+  const claimsSheet = financeGSheet.sheetsByTitle['Claims'];
+  const allRecordSheet = financeGSheet.sheetsByTitle['All Records'];
+  const reimbursementSheet = financeGSheet.sheetsByTitle['Reimbursement'];
+
+  await claimsSheet.loadCells();
+  await allRecordSheet.loadCells();
+  await reimbursementSheet.loadCells();
+
+  const claimRowNo = await searchRowNo(
+    callback,
+    claimsSheet,
+    'A',
+    2,
+    claimsSheet.rowCount
+  );
+  const allRecordRowNo = await searchRowNo(
+    callback,
+    allRecordSheet,
+    'B',
+    2,
+    allRecordSheet.rowCount
+  );
+  const reimbursementRowNo = await searchRowNo(
+    callback,
+    reimbursementSheet,
+    'B',
+    2,
+    reimbursementSheet.rowCount
+  );
+  const claimRows = await claimsSheet.getRows({});
+  const allRecordRows = await allRecordSheet.getRows({});
+  const reimbursementRows = await reimbursementSheet.getRows({});
+  let claimRow: GoogleSpreadsheetRow | null = null;
+  let allRecordRow: GoogleSpreadsheetRow | null = null;
+  let reimbursementRow: GoogleSpreadsheetRow | null = null;
+
+  if (
+    claimRowNo != -1 &&
+    claimRowNo &&
+    allRecordRowNo != -1 &&
+    allRecordRowNo &&
+    reimbursementRowNo != -1 &&
+    reimbursementRowNo
+  ) {
+    claimRow = claimRows[claimRowNo - 2];
+    allRecordRow = allRecordRows[allRecordRowNo - 2];
+    reimbursementRow = reimbursementRows[reimbursementRowNo - 2];
+  } else {
+    await ctx.reply('Invalid Claim');
+    return;
+  }
+
+  if (claimRow && allRecordRow && reimbursementRow) {
+    await claimRow.delete();
+    await allRecordRow.delete();
+    await reimbursementRow.delete();
+  } else {
+    await ctx.reply('Invalid Claim');
+    return;
+  }
+
+  await Database.getMongoRepository(Claims).deleteOne({
+    claimid: callback,
+  });
+  await ctx.reply('Claim Deleted');
 };
 
 const changeFinancePassword = async (ctx: CallbackQueryContext<BotContext>) => {
